Optimize filterbank, FFT, channels, bitstream, and huffman modules
- Refactor `filtbank.c` windowing with C11 inline functions and restrict pointers, eliminating struct allocations while retaining `CopyFlat`/`ZeroFlat`.
- Optimize `channels.c` short-window ICS grouping using direct bitwise shift masks.
- Streamline `bitstream.c` `PutBit()` fast-path byte calculations and remove unused `ByteAlign()` dead code.
- Pre-compute Radix-4 DIF twiddles and reorder tables during `fft_initialize()` in `fft.c`.
- Refactor `huff2.c` spectral coding and inline escape helper using C11 restrict pointers.
- Add integer cast bounds safety in `util.c` MaxBitrate calculation.

+++ libfaac/util.c
@@ -1,6 +1,7 @@
 /*
  * FAAC - Freeware Advanced Audio Coder
  * Copyright (C) 2001 Menno Bakker
+ * Copyright (C) 2026 Nils Schimmelmann
  *
  * This library is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
@@ -42,7 +43,7 @@ int GetSRIndex(unsigned int sampleRate)
 unsigned int MaxBitrate(unsigned long sampleRate)
 {
     /* ISO/IEC 14496-3 maximum frame payload: 6144 bits per channel */
-    return AAC_MAX_BITS_PER_CH * sampleRate / FRAME_LEN;
+    return (unsigned int)((uint64_t)AAC_MAX_BITS_PER_CH * sampleRate / FRAME_LEN);
 }
 
 /* Portable CLZ (returns 32 for x==0); lets escape() get a magnitude's bit-length
@@ -53,7 +54,7 @@ int CountLeadingZeros(unsigned int x)
 #ifdef _MSC_VER
     unsigned long leading_zero;
     _BitScanReverse(&leading_zero, x);
-    return 31 - leading_zero;
+    return 31 - (int)leading_zero;
 #elif defined(__GNUC__) || defined(__clang__)
     return __builtin_clz(x);
 #else
@@ -66,4 +67,3 @@ int CountLeadingZeros(unsigned int x)
     return n;
 #endif
 }
-
